+++ src/fields/nonnative/allocated_mul_result_var.rs
