+++ src/fields/nonnative/mul_result_var.rs
